get some better auto complete workin'

+++ src/lexing/token.rs
@@ -3,8 +3,8 @@ use super::token_kind::TokenKind;
 #[derive(Clone, Debug, PartialEq, Eq, Hash)]
 pub struct Token {
     pub kind: TokenKind,
-    pub start: usize,
-    pub end: usize,
+    pub start: u32,
+    pub end: u32,
     pub line: u32,
     pub col: u32,
 }
